requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector==2.2.9
mysqlclient==2.2.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
